Implement a basic blacklist
One of the problems with a simple Q learner is that it assumes that
the functions (which represent state) always behave the same in terms
of the characters they compare. However, functions like
has_char(char_val) changes the comparisons they make depending on the
input. This is similar to the grammar learner. We make a beginning to
fixing it by looking for comparisons made on entering a state, and
if the comparisons are not similar to previous comparisons in the
state, mark the state as a changeling under the blacklist.

Note that there are additional complications here. In particular due
to the presence of loops, we can not really do substring string
matches. Hence, for now we use set comparisons to identify changelings.
This should be revisited when basic block ids are incorporated.

For these changelings, we simply use random choice for now. In futre
we will implement a Q table exclusively for changelings, with state
identified with function name + (possibly id) + comparisons made