Changes: Use the blockid to distinguish comparisons in the same function
Previously, we used the function name as the name of the state. However,
this is underpowered as a function may not directly correspond to a node
in an automata.

A better state is a basic block in a function (an even better one would
be the precise comparison point). We use the functionname@basicblockid
as the new name of the state.